Fix libssl-dev dependency for Ubuntu 18.04
46b6bde5 seemed to focus on fixing a dependency for Debian 10, however
in doing so it swapped the libssl-dev dependency for Ubuntu 18.04 to be
libssl1.0-dev. Ruby 2.4+ is compatible with OpenSSL 1.1, so you'd only
need OpenSSL 1.0 if you were installing Ruby 2.3 or older, none of which
are currently supported by Ruby.

This regression created a conflict with libmysqlclient-dev as it depends
on libssl-dev (at least as of v5.7.31), so if you install
libmysqlcient-dev first, it'll install libssl-dev, then if you use this
cookbook, it'll install libssl1.0-dev, which conflicts with libssl-dev,
thus removing libssl-dev and libmysqlclient-dev, breaking anything that
tries to load the MySQL development files, such as the mysql2 gem.